+++ tests/test_coverage.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 from iranges import IRanges
 
 __author__ = "jkanche"
@@ -13,26 +12,75 @@ def test_coverage():
     x = IRanges(starts, widths)
 
     cov = x.coverage()
-    assert all(np.equal(cov, [1,1,1,1,1,1,1,2,2,2,2,2,1,1,2,2,2,2,3,2,2,2,2,2,1,1,1,1,1,0,0,0,0,1,1,0,0,0,0,1,1,1,1,1,1,1]))
+    assert all(
+        np.equal(
+            cov,
+            [
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                2,
+                2,
+                2,
+                2,
+                2,
+                1,
+                1,
+                2,
+                2,
+                2,
+                2,
+                3,
+                2,
+                2,
+                2,
+                2,
+                2,
+                1,
+                1,
+                1,
+                1,
+                1,
+                0,
+                0,
+                0,
+                0,
+                1,
+                1,
+                0,
+                0,
+                0,
+                0,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+                1,
+            ],
+        )
+    )
 
     starts = [-2, 6, 9, -4, 1, 0, -6, 10]
     widths = [5, 0, 6, 1, 4, 3, 2, 3]
     x = IRanges(starts, widths)
 
     cov = x.coverage()
-    assert all(
-        np.equal(cov, [3,3,1,1,0,0,0,0,1,2,2,2,1,1])
-    )
+    assert all(np.equal(cov, [3, 3, 1, 1, 0, 0, 0, 0, 1, 2, 2, 2, 1, 1]))
+
 
 def test_coverage_with_shift():
     starts = [-2, 6, 9, -4, 1, 0, -6, 10]
     widths = [5, 0, 6, 1, 4, 3, 2, 3]
     x = IRanges(starts, widths)
 
-    cov = x.coverage(shift = 7)
-    assert all(
-        np.equal(cov, [1,1,1,0,1,1,2,3,3,1,1,0,0,0,0,1,2,2,2,1,1])
-    )
+    cov = x.coverage(shift=7)
+    assert all(np.equal(cov, [1, 1, 1, 0, 1, 1, 2, 3, 3, 1, 1, 0, 0, 0, 0, 1, 2, 2, 2, 1, 1]))
 
 
 def test_coverage_with_shift_and_width():
@@ -41,9 +89,7 @@ def test_coverage_with_shift_and_width():
     x = IRanges(starts, widths)
 
     cov = x.coverage(shift=7, width=27)
-    assert all(
-        np.equal(cov, [1,1,1,0,1,1,2,3,3,1,1,0,0,0,0,1,2,2,2,1,1,0,0,0,0,0,0])
-    )
+    assert all(np.equal(cov, [1, 1, 1, 0, 1, 1, 2, 3, 3, 1, 1, 0, 0, 0, 0, 1, 2, 2, 2, 1, 1, 0, 0, 0, 0, 0, 0]))
 
 
 def test_coverage_with_weight():
@@ -52,6 +98,4 @@ def test_coverage_with_weight():
     x = IRanges(starts, widths)
 
     cov = x.coverage(weight=10)
-    assert all(
-        np.equal(cov, [30,30,10,10,0,0,0,0,10,20,20,20,10,10])
-    )
+    assert all(np.equal(cov, [30, 30, 10, 10, 0, 0, 0, 0, 10, 20, 20, 20, 10, 10]))