+++ tests/test_order_ops.py
@@ -1,5 +1,4 @@
 import numpy as np
-import pytest
 from iranges import IRanges
 
 __author__ = "jkanche"