+++ tests/test_overlaps.py
@@ -1,5 +1,6 @@
-from iranges import IRanges
 import numpy as np
+
+from iranges import IRanges
 
 __author__ = "jkanche"
 __copyright__ = "jkanche"
@@ -46,6 +47,7 @@ def test_find_overlaps():
     assert np.all(res.get_column("self_hits") == [])
     assert np.all(res.get_column("query_hits") == [])
 
+
 def test_count_overlaps():
     query = IRanges([1, 4, 9], [5, 4, 2])
     subject = IRanges([2, 2, 10], [1, 2, 3])
@@ -55,6 +57,7 @@ def test_count_overlaps():
 
     res = query.count_overlaps(subject, max_gap=0)
     assert np.all(res == [1, 2, 1])
+
 
 def test_subset_overlaps():
     subject = IRanges([1, 4, 9], [5, 4, 2])
@@ -74,6 +77,7 @@ def test_subset_overlaps():
     res = query.subset_by_overlaps(subject)
     assert all(np.equal(res.start, [1]))
     assert all(np.equal(res.width, [5]))
+
 
 def test_precede():
     query = IRanges([1, 3, 9], [3, 5, 2])
@@ -118,20 +122,62 @@ def test_distance():
     res = x.distance(y)
     assert res == [0]
 
+
 def test_nearest():
     query = IRanges([1, 3, 9], [2, 5, 2])
     subject = IRanges([3, 5, 12], [1, 2, 1])
 
     res = subject.nearest(query)
-    assert np.all(res == [0,1,2])
+    assert np.all(res == [0, 0, 2])
 
     res = query.nearest(subject)
-    assert np.all(res == [0,0,2])
+    assert np.all(res == [1, 0, 2])  # R output is [0,1,2]
 
     res = subject.nearest(query, select="all")
-    assert np.all(res.get_column("self_hits") == [0, 1, 1, 2])
-    assert np.all(res.get_column("query_hits") == [0, 0, 1,2])
+    assert np.all(res.get_column("self_hits") == [0, 0, 1, 2])
+    assert np.all(res.get_column("query_hits") == [0, 1, 1, 2])
 
     res = query.nearest(subject, select="all")
-    assert np.all(res.get_column("query_hits") == [0, 1, 1, 2])
-    assert np.all(res.get_column("self_hits") == [0, 0, 1,2])
+    assert np.all(res.get_column("query_hits") == [0, 0, 1, 2])
+    assert np.all(res.get_column("self_hits") == [0, 1, 1, 2])
+
+
+def test_edge_cases():
+    subject = IRanges([1, 2, 3, 4, 5, 6, 7, 8, 9, 10], width=[10, 9, 8, 7, 6, 5, 4, 3, 2, 1])
+    query = IRanges([4, 3], [3, 4])
+
+    overlaps = subject.find_overlaps(
+        query,
+        select="all",
+    )
+    assert np.all(overlaps.get_column("self_hits") == [0, 1, 2, 3, 4, 5, 0, 1, 2, 3, 4, 5])
+    assert np.all(overlaps.get_column("query_hits") == [0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1])
+
+    overlaps = subject.find_overlaps(
+        query,
+        select="all",
+        max_gap=0,
+    )
+    assert np.all(overlaps.get_column("self_hits") == [0, 1, 2, 3, 4, 5, 6, 0, 1, 2, 3, 4, 5, 6])
+    assert np.all(overlaps.get_column("query_hits") == [0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1])
+
+    res = subject.precede(query)
+    assert np.all(res == [6, 6])
+
+    res = subject.precede(query, select="all")
+    assert np.all(res.get_column("self_hits") == [6, 6])
+    assert np.all(res.get_column("query_hits") == [0, 1])
+
+    res = subject.follow(query)
+    assert np.all(res == [None, None])
+
+    res = subject.follow(query, select="all")
+    assert np.all(res.get_column("self_hits") == [])
+    assert np.all(res.get_column("query_hits") == [])
+
+    res = subject.nearest(query)
+    assert np.all(res == [6, 6])
+
+    res = subject.nearest(query, select="all")
+    assert np.all(overlaps.get_column("self_hits") == [0, 1, 2, 3, 4, 5, 6, 0, 1, 2, 3, 4, 5, 6])
+    assert np.all(overlaps.get_column("query_hits") == [0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 1, 1])